Smarter frontmatter

+++ .github/workflows/scripts/download_readmes.sh
@@ -6,22 +6,38 @@
 jq -c -r '( .NAME + " " + .REPO + " " + .URL + " " + .LANG + " " + .DATE + " " + .READ )' ./repos.json | while read name repo url lang date readme; do
 
     echo "${name}"
-    ## Make a directory
-    mkdir -p ${name}
+
+    # Skip if already exists
+    if [[ ! -f "./${name}" ]]; then
+        ## Make a directory
+        mkdir -p ${name}
+    fi
     
-    ## Download readme as index.mdx
     cd ${name}
-    /usr/bin/curl -qs "${readme}" -O readme.md
+
+    ## File doesn't exist, so create it.
+    if [[ ! -f "./index.mdx" ]]; then
     
-    ## Prefix Frontmatter
-    echo "---" > index.mdx
-    echo "slug:  \"/projects/${name}\"" >> index.mdx
-    echo "date:  \"$date\"" >> index.mdx
-    echo "title: \"${name}\"" >> index.mdx
-    echo "icon:  \"$lang\"" >> index.mdx
-    echo "---" >> index.mdx
-    printf "\n\n" >> index.mdx
-    cat README.md >> index.mdx
+        /usr/bin/curl -qs "${readme}" -O readme.md
+        
+        ## Prefix Frontmatter
+        echo "---" > index.mdx
+        echo "slug:  \"/projects/${name}\"" >> index.mdx
+        echo "date:  \"$date\"" >> index.mdx
+        echo "title: \"${name}\"" >> index.mdx
+        echo "icon:  \"$lang\"" >> index.mdx
+        echo "---" >> index.mdx
+        printf "\n\n" >> index.mdx
+        cat README.md >> index.mdx
+
+    else
+        # Keep Frontmatter already set. 
+        # May have been manually changed, so don't alter it.
+        head -n 7 index.mdx > frontmatter.md && mv frontmatter.md index.mdx
+        /usr/bin/curl -qs "${readme}" -O readme.md
+        cat README.md >> index.mdx
+
+    fi
 
     # Clean up Readme.
     rm README.md
